chore: Add Styles class
