+++ src/components/Text/Text.test.tsx
@@ -2,11 +2,8 @@ import React from 'react';
 
 import { render, screen } from '@testing-library/react-native';
 import { Text } from './Text';
-import {
-  BrandConfig,
-  BrandConfigProvider,
-} from '../components/BrandConfigProvider';
-import { colors } from '../themes/default';
+import { BrandConfig, BrandConfigProvider } from '../BrandConfigProvider';
+import { colors } from '../../themes/default';
 
 describe('Without configuration', () => {
   test('Renders text with default style (baseTheme.colors.text)', () => {
